Add very basic dependency support
Load all items in the right order, assuming that all required
dependencies are selected. Don't try to find out which dependencies are
needed but not checked, and which ones aren't installed.
Don't check whether there are mods installed that have the same 'name',
which would also mess up with the dependency support.